26.02.22

+++ Accountant/Dtos/ProfileAccountDto.cs
@@ -4,11 +4,11 @@ public class ProfileAccountDto
 {
     public int ProfileId { get; set; }
     public int AccountId { get; set; }
-    public double ShareOwned { get; set; }
+    public decimal ShareOwned { get; set; }
 }
 
 public class ChangeProfileAccountDto
  {
      public int Id { get; set; }
-     public double ShareOwned { get; set; }
+     public decimal ShareOwned { get; set; }
  }
